requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
redis==7.1.0
sqlparse==0.5.5
gunicorn>=21.2
psycopg[binary]==3.3.2
python-dotenv==1.0.1
